Initialize CNTVOFF_EL2 to zero across all CPUs
The CNTVOFF register value is UNKNOWN after reset and shall need
initialization by the firmware or the operating system hypervisor.
The Generic Timer always includes a virtual counter. The virtual
counter contains the value of the physical counter minus a 64-bit
virtual offset. When executing in a Non-secure EL1 or EL0 mode, the
virtual offset value relates to the current virtual machine.

 * The CNTVOFF register contains the virtual offset that defines the
   virtual offset between the physical count and the virtual count.
 * CNTVOFF is only accessible from EL2, or from EL3 when SCR.NS is
   set to 1.
 * The CNTVOFF register value is UNKNOWN after reset. If EL2 is not
   implemented, the virtual counter uses a fixed virtual offset of
   zero.

Operating systems expect this value to be initialized to zero or any
other value which is same across all processors, to have a consistent
view. Since the operating system running at EL1 does not have access
to this register, the CNTVOFF needs to be set to zero by the firmware
either in EL3 or EL2.

The following scenarios describe considerations for which Execution
Level CNTVOFF may be initialized.

Scenario 1: AArch32 Implementation that includes EL3 but not EL2
  CNTVOFF in this implementation is not supported.  The behavior of
  accesses to CNTVOFF is as follows:
  * Any access from Secure Monitor mode is UNPREDICTABLE, regardless
    of the value of SCR.NS.
  * All other accesses are UNDEFINED.

Scenario 2: Implementation with no EL2
  CNTVOFF is set to 0 as defined by architecture. So no change needed
  in firmware.

Scenario 3: Implementation that includes EL3 and EL2 + MP Parking
  Protocol
  The CNTVOFF may be setup by UEFI running in EL2 when SCR.NS = 0 as
  well as in secure firmware.

Scenario 4: Implementation that includes EL3 and EL2 + PSCI
  The CNTVOFF for the primary CPU can be setup by UEFI running in EL2,
  and the Hypervisor code sets this up for the secondary CPUs.  Or it
  can be done in secure firmware.

ARM Trusted firmware covers the scenarios listed above. However some
operating systems hypervisors (Scenario 4) expect that the CNTVOFF is
configured by the firmware. Their argument being that there is no
clear boundary defined by any specification as to where the CNTVOFF
should be configured.

Since ARM Trusted firmware is configuring the CNTVOFF in some scenarios,
it will be consistent to handle the case mentioned above. This is purely
to say that CNTVOFF initialization is handled at only one place.

Whether this should be done in the firmware or the hypervisor is a matter
of debate and clarification should be provided by the firmware
specifications.

Under such circumstances this fix should be considered as a workaround.

Change-Id: I2d874fd5d3a40cba6882daa9f91167160f007c3d
Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>