fiptool: Add Visual Studio project
Add the project files required to generate a Visual Studio build of
Fiptool.  This can be used to generate the tool in a windows environment
This commit does not enable use of make on Windows.
Howevr, once fiptool.exe has been generated it can be used to satisfy
the make build requirements.

Change-Id: Ida2894807dbbf8171a64e2763250977a2c010e06
Signed-off-by: Evan Lloyd <evan.lloyd@arm.com>